requirements.txt
Flask
python-dotenv
google-genai
Pillow
arabic-reshaper
gunicorn
